Fixes to overall system
