+++ src/backend/apps/entity_identification/main.py
@@ -8,6 +8,7 @@ import requests
 import sentence_ident as si
 import article_ident as ai
 
+
 app = Flask(__name__)
