+++ src/backend/apps/prediction/main.py
@@ -6,6 +6,7 @@ import json
 import prediction
 import sentence_to_points
 
+
 app = Flask(__name__)
 
 
@@ -26,19 +27,11 @@ def prediction_for_company():
 @app.route('/points/sentences', methods=['POST'])
 def sentence_points():
     sentence = request.get_json()
+    print(sentence)
     sentence_to_points.sentence_to_point(sentence)
     return flask.Response(status=200)
+
 
 if __name__ == '__main__':
     app.run(port=5004)
 
-
-# Updates verdict of company in database
-#def update_verdict(stock_code, verdict, preds):
-#    url = baseurl + '/company/predictions'
-#    payload = {'stock_code': stock_code, 'verdict': verdict, 'predictions': json.dumps(preds)}
-#    r = requests.post(url, json=payload)
-#    if r.status_code == 200:
-#        return "verdict updated"
-#    else:
-#        return "something went wrong"
