+++ src/backend/apps/prediction/sentence_to_points.py
@@ -28,16 +28,21 @@ def sentence_to_point(sentence_json):
     url = baseurl + "/points"
     date_time = sentence_json['date'] + " " + sentence_json['time']
     data = stock_data.date_at_date(sentence_json['context'], sentence_json['date'] + " " + sentence_json['time'])
-    payload = {"time": date_time,
-               "company_id": sentence_json['context'],
-               "sentiment": sentence_json['sentiment'],
-               "sentence_id": sentence_json['id'],
-               "open": data['Open'],
-               "high": data['High'],
-               "low": data['Low'],
-               "close": data['Close'],
-               "volume": data['Volume'], }
-    r = requests.post(url, json=payload)
+    print(data)
+    if data.empty:
+        payload = {"sentence_id": sentence_json['id'], "close": None}
+        r = requests.post(url, json=payload)
+    else:
+        payload = {"time": date_time,
+                   "company_id": sentence_json['context'],
+                   "sentiment": sentence_json['sentiment'],
+                   "sentence_id": sentence_json['id'],
+                   "open": data['Open'],
+                   "high": data['High'],
+                   "low": data['Low'],
+                   "close": data['Close'],
+                   "volume": data['Volume'], }
+        r = requests.post(url, json=payload)
 
 
 # Update stock data of point
