+++ src/backend/backend_api/app.py
@@ -421,13 +421,20 @@ def get_data_points(company_id):
 def add_data_point():
     if request.is_json:
         try:
+            content = request.get_json()
+            if content['close'] == None:
+                sentence = Sentence.query.get(content['sentence_id'])
+                sentence.status = "BLOCKED"
+                db.session.commit()
+                return flask.Response(status=204)
+
             conn = psycopg2.connect(host='localhost',
                                     port='5433',
                                     user='postgres',
                                     password='2206',
                                     database='company_data')
             cursor = conn.cursor()
-            content = request.get_json()
+
 
             datetime_obj = datetime.strptime(content['time'], "%m/%d/%Y %H:%M:%S")
             cursor.execute("INSERT INTO points(time, company_id, sentiment, sentence_id, open, high, low, close, volume) "
